requirements.txt
jupyter
jupyterlab
flask
numpy
pandas
toyplot
requests
calysto
matplotlib
scipy
seaborn
ipywidgets
sql
sqlalchemy
ipython-sql

